feat(sampler): 動的バッチング用の音声長ファイル作成スクリプトを追加 refactor(config): 設定ファイルに長さファイルパスとバッチビン数を追加 refactor(dataset): データセット作成時に長さファイルを読み込むように変更 refactor(train): DataLoader作成時にLengthBatchSamplerの設定を改善 refactor(model): 損失計算時のフレーム数の最大差チェックを修正 style(docs): ドキュメントの表現を改善

+++ src/unofficial_slash/network/dsp/vuv_detector.py
@@ -50,7 +50,7 @@ def voicing_detection(
     v_continuous = Mp_safe / (Mp_safe + Map_safe)  # (B, T)
 
     # 二値のV/UV判定: v = 1 if v' ≥ θ else 0
-    v_binary = (v_continuous >= threshold).float()  # (B, T)
+    v_binary = v_continuous >= threshold  # (B, T)
 
     return v_continuous, v_binary
